split throttling test to node
